moved all common initialization to TestUtils
